feat: add health option to jx diagnose
